requirements.txt
absl-py==0.9.0
appdirs==1.4.3
asn1crypto==1.2.0
astor==0.8.1
certifi==2019.11.28
cffi==1.13.0
chardet==3.0.4
conda==4.8.1
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
distlib==0.3.0
edward==1.3.5
filelock==3.0.12
gast==0.3.2
grpcio==1.26.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
numpy==1.18.1
opt-einsum==3.2.0
pandas==0.25.3
patsy==0.5.1
protobuf==3.11.2
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyro-api==0.1.1
pyro-ppl==1.3.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
ruamel-yaml==0.15.46
scikit-learn==0.22.2.post1
scipy==1.2.0
seaborn==0.9.0
six==1.12.0
statsmodels==0.10.1
tensorboard==1.12.2
tensorflow==1.13.0rc1
tensorflow-estimator==1.13.0
termcolor==1.1.0
torch==1.4.0
tornado==6.0.3
tqdm==4.36.1
urllib3==1.24.2
virtualenv==20.0.14
Werkzeug==0.16.0
zipp==3.1.0
